refactor: home / hero-section
